+++ confik/src/third_party.rs
@@ -119,6 +119,62 @@ mod ipnetwork {
     }
 }
 
+#[cfg(feature = "js_option")]
+mod js_option {
+    use js_option::JsOption;
+    use serde::de::DeserializeOwned;
+
+    use crate::{Configuration, ConfigurationBuilder};
+
+    impl<T> Configuration for JsOption<T>
+    where
+        T: DeserializeOwned + Configuration,
+    {
+        type Builder = JsOption<<T as Configuration>::Builder>;
+    }
+
+    impl<T> ConfigurationBuilder for JsOption<T>
+    where
+        T: DeserializeOwned + ConfigurationBuilder,
+    {
+        type Target = JsOption<<T as ConfigurationBuilder>::Target>;
+
+        fn merge(self, other: Self) -> Self {
+            match (self, other) {
+                // If both `Some` then merge the contained builders
+                (Self::Some(us), Self::Some(other)) => Self::Some(us.merge(other)),
+                // If we don't have a value then always take the other
+                (Self::Undefined, other) => other,
+                // Either:
+                // - We're explicitly `Null`
+                // - We're explicitly `Some` and the other is `Undefined` or `Null`
+                //
+                // In either case, just take our value, which should be preferred to other.
+                (us, _) => us,
+            }
+        }
+
+        fn try_build(self) -> Result<Self::Target, crate::Error> {
+            match self {
+                Self::Undefined => Ok(Self::Target::Undefined),
+                Self::Null => Ok(Self::Target::Null),
+                Self::Some(val) => Ok(Self::Target::Some(val.try_build()?)),
+            }
+        }
+
+        fn contains_non_secret_data(&self) -> Result<bool, crate::UnexpectedSecret> {
+            match self {
+                Self::Some(data) => data.contains_non_secret_data(),
+
+                // An explicit `Null` is counted as data, overriding any default.
+                Self::Null => Ok(true),
+
+                Self::Undefined => Ok(false),
+            }
+        }
+    }
+}
+
 #[cfg(feature = "secrecy")]
 mod secrecy {
     use secrecy::SecretString;