+++ confik/tests/third_party.rs
@@ -126,3 +126,76 @@ mod bigdecimal {
         }
     }
 }
+
+#[cfg(feature = "js_option")]
+mod js_option {
+    use confik::{Configuration, TomlSource};
+    use js_option::JsOption;
+
+    #[derive(Configuration, Debug)]
+    struct Config {
+        opt: JsOption<usize>,
+    }
+
+    #[test]
+    fn undefined() {
+        let config = Config::builder()
+            .try_build()
+            .expect("Should be valid without config");
+        assert_eq!(config.opt, JsOption::Undefined);
+    }
+
+    #[cfg(feature = "json")]
+    #[test]
+    fn null() {
+        let json = r#"{ "opt": null }"#;
+
+        let config = Config::builder()
+            .override_with(confik::JsonSource::new(json))
+            .try_build()
+            .expect("Failed to parse config");
+        assert_eq!(config.opt, JsOption::Null);
+    }
+
+    #[test]
+    fn present() {
+        let toml = "opt = 5";
+
+        let config = Config::builder()
+            .override_with(TomlSource::new(toml))
+            .try_build()
+            .expect("Should be valid without config");
+        assert_eq!(config.opt, JsOption::Some(5));
+    }
+
+    #[cfg(feature = "json")]
+    #[test]
+    fn merge() {
+        #[derive(Debug, Configuration, PartialEq, Eq)]
+        struct Config {
+            one: JsOption<usize>,
+            two: JsOption<usize>,
+            three: JsOption<usize>,
+            four: JsOption<usize>,
+        }
+
+        let base = r#"{ "two": null, "three": 5 }"#;
+        let merge = r#"{ "one": 1, "two": 2, "three": 3}"#;
+
+        let config = Config::builder()
+            .override_with(confik::JsonSource::new(merge))
+            .override_with(confik::JsonSource::new(base))
+            .try_build()
+            .expect("Failed to parse config");
+
+        assert_eq!(
+            config,
+            Config {
+                one: JsOption::Some(1),
+                two: JsOption::Null,
+                three: JsOption::Some(5),
+                four: JsOption::Undefined,
+            }
+        );
+    }
+}